moved force daily reset to it's own preference area
